% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jaspModuleTools.R
\name{compile}
\alias{compile}
\title{Compiles given jaspModule and (optionally) creates a JaspModuleBundle}
\usage{
jaspModuleTools::updateLockfile('~/jaspTTest/')
}
\arguments{
\item{moduledir}{Path to the jaspModule root folder}

\item{workdir}{Path where the renv cache and pkglib will be created}

\item{resultdir}{Path to dir where the resulting bundle will be created}

\item{createBundle}{Should a JASP bundle be created}

\item{bundleAll}{Should all dependencies be bundled? by default we bundle only those not present in the Remote Cellar Repo}

\item{buildforJaspVersion}{Specifies the jaspVersion for which this module should be compiled and bundeld}

\item{useJASPRemoteCellar}{should we use the Remote Cellar?}

\item{repoName}{which remote repo should be used? for expample 0.19.3 or development}

\item{localCellar}{local cellar path to also include for development purposes}
}
\description{
Compiles given jaspModule and (optionally) creates a JaspModuleBundle. Will gather binaries from the JASP Remote Cellar Repo.
}
