% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jaspModuleTools.R
\name{updateLockfile}
\alias{updateLockfile}
\title{Updates lockfile using pkgdepends}
\usage{
jaspModuleTools::updateLockfile('~/jaspTTest/')
}
\arguments{
\item{moduledir}{Path to the jaspModule root folder}
}
\description{
Updates the lockfile of a given jaspModule using pkgdepens
}
